--- Assets/_InDevelopment/Chase/InProgressScripts/AnimatorColorerStateListener.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 1a4d045c108242409f4db5e9d626ad65
-timeCreated: 1717681844
