+++ Assets/_InDevelopment/Chase/InProgressScripts/EventButton.cs
@@ -1,7 +1,5 @@
-﻿using System;
-using UnityEngine;
+﻿using UnityEngine;
 using UnityEngine.Events;
-using UnityEngine.EventSystems;
 
 public class EventButton : UIButton
 {
