+++ Assets/_InDevelopment/Chase/InProgressScripts/UIButton.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using UnityEngine;
 using UnityEngine.UI;
-using UnityEngine.EventSystems;
 
 [RequireComponent(typeof(Button))]
 public abstract class UIButton : MonoBehaviour, IButton
